backend/requirements.txt
# Core Framework
fastapi==0.103.2
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database (SQLite - async support)
sqlalchemy==2.0.44
alembic==1.13.0
aiosqlite==0.19.0

# Validation (updated to 2.x)
pydantic==2.12.3
pydantic-settings==2.5.2

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# AI APIs (Python 3.14 compatible versions)
anthropic>=0.72.0
openai==1.10.0
google-generativeai>=0.8.0

# Utilities
python-dotenv==1.0.0
email-validator==2.1.0
requests==2.31.0
httpx==0.25.2

# Web Scraping
beautifulsoup4==4.12.2
lxml==5.1.0

# System Monitoring & Management
psutil==5.9.8
schedule==1.2.0

# Document Generation (Blog Export)
python-docx==1.1.0
Pillow==10.1.0  # Required for image handling in python-docx
